refactor(objects): carve file-size headroom ahead of the alarm/event batches
No behavior change. Pure moves, verified by the suite: 2234 tests, exit 0.

crates/bacnet-objects/src/multistate/mod.rs sat at exactly 700/700, so a single
added line failed CI — and #123 has to touch it. Six more files in this crate
were within 35 lines of the cap, which the upcoming event-state batches would
have breached mid-PR.

Two shapes, both following patterns already in the crate:

- Split multi-object families one file per object type, as access_control/
  already does (door.rs, credential.rs, zone.rs, …). multistate/ and analog/
  become mod.rs + input.rs + output.rs + value.rs, with the shared header and
  require_nonzero_states staying in mod.rs.
- Extract inline `mod tests { … }` bodies to a sibling module, as event.rs +
  event/tests.rs already does.

  multistate/mod.rs                700 -> 23   (+204/242/240)
  analog/mod.rs                    685 -> 15   (+202/235/242)
  notification_class/tests/mod.rs  693 -> 27   (+301/310/66)
  life_safety.rs                   688 -> 341  (+346 tests)
  elevator.rs                      688 -> 452  (+235 tests)
  network_port.rs                  675 -> 230  (+444 tests)
  event_log.rs                     463 -> 211  (+251 tests)

Largest file in the crate goes 700 -> 669.

elevator.rs and network_port.rs were not on the plan's list; they measured 688
and 675, so they are included rather than left to breach later.

make_time and make_dest_device stay in notification_class/tests/mod.rs because
recipient_list.rs already imports them via `use super::make_time`; siblings pull
them the same way.

Public paths are unchanged — each split family re-exports with `pub use`, so
`bacnet_objects::multistate::MultiStateInputObject` still resolves. No CHANGELOG
entry: nothing user-visible changed.
